Split tmux session bottom controls
